requirements.txt
Django>=1.6
South
django-oauth-toolkit
django-annoying
django-tastypie
